Do not show NFC fragment on challenge-response
There were two operations on login previously: login and sign for
challenge-response. Nfc fragment is shown on every token operation
and in this case it was showed twice. These operations are now very
similar to payment signing because it performs the same pkcs11 calls.
So we decided to create one token method for them.