Ignore errors on certificate reading
Before this commit certificate reading errors led to the fact that it
would be impossible to work with the token at all. But some cases are
very common. For example, you could have your CA certificate on token
without keys for it or you have not GOST certificates on token.